Split tree.py into its own subpackage
