+++ src/tqec/compile/tree/__init__.py
@@ -1,0 +1,1 @@
+from .tree import LayerTree as LayerTree
